User 아키텍처 구조 변경
