+++ src/main/java/kr/hhplus/be/server/config/jpa/JpaConfig.java
@@ -12,7 +12,7 @@ import org.springframework.transaction.PlatformTransactionManager;
 @EnableJpaRepositories(
         basePackages = {
             "kr.hhplus.be.server.product.repository",
-            "kr.hhplus.be.server.user.adapter.gateway",
+                "kr.hhplus.be.server.user.infra.gateway",
             "kr.hhplus.be.server.order.adapter.gateway"
         }
 )
