--- src/main/java/kr/hhplus/be/server/user/adapter/gateway/UserJpaRepo.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.user.adapter.gateway;
-
-import kr.hhplus.be.server.user.adapter.gateway.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserJpaRepo extends JpaRepository<UserEntity, Long>, UserRepositoryCustom {
-    Optional<UserEntity> findByEmail(String email);
-}
