--- src/main/java/kr/hhplus/be/server/user/adapter/gateway/UserRepositoryCustom.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.user.adapter.gateway;
-
-public interface UserRepositoryCustom {
-
-}
