--- src/main/java/kr/hhplus/be/server/user/adapter/gateway/UserRepositoryImpl.java
@@ -1,13 +1,0 @@
-package kr.hhplus.be.server.user.adapter.gateway;
-
-import jakarta.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class UserRepositoryImpl implements UserRepositoryCustom {
-
-    private final EntityManager em;
-
-}
