+++ src/main/java/kr/hhplus/be/server/user/application/usecase/command/UserChargeBalanceCommand.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.user.application.usecase.command;
+
+public record UserChargeBalanceCommand(
+        Long userId,
+        int amount
+) {
+}
