+++ src/main/java/kr/hhplus/be/server/user/infra/gateway/UserJpaRepo.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.user.infra.gateway;
+
+import kr.hhplus.be.server.user.infra.gateway.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserJpaRepo extends JpaRepository<UserEntity, Long>, UserRepositoryCustom {
+    Optional<UserEntity> findByEmail(String email);
+}
