+++ src/main/java/kr/hhplus/be/server/user/infra/gateway/UserRepositoryCustom.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.user.infra.gateway;
+
+public interface UserRepositoryCustom {
+
+}
