+++ src/main/java/kr/hhplus/be/server/user/infra/gateway/UserRepositoryCustomImpl.java
@@ -1,0 +1,13 @@
+package kr.hhplus.be.server.user.infra.gateway;
+
+import jakarta.persistence.EntityManager;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class UserRepositoryCustomImpl implements UserRepositoryCustom {
+
+    private final EntityManager em;
+
+}
