--- src/main/java/kr/hhplus/be/server/user/usecase/in/UserChargeBalanceCommand.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.user.usecase.in;
-
-public record UserChargeBalanceCommand(
-        Long userId,
-        int amount
-) {
-}
